unify android and ios files (there is only one difference)
